apps/user_service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "common",
    "fastapi >=0.104.0",
    "uvicorn[standard] >=0.24.0",
    "python-jose[cryptography] >=3.3.0",
    "passlib[bcrypt] >=1.7.4",
    "sqlalchemy >=2.0.0",
    "asyncpg >=0.29.0",
    "httpx >=0.25.0",
    "pytest >=7.4.0",
    "pytest-asyncio >=0.21.0",
    "pytest-cov >=4.1.0",
    "black >=23.0.0",
    "ruff >=0.1.0",
    "pyjwt==2.10.1",
    "email-validator >=2.0.0"
]

[tool.hatch.build]
include = ["src/user_service"]

[tool.hatch.build.targets.wheel]
packages = ["src/user_service"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.coverage.run]
source = ["src/user_service"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/migrations/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == '__main__':",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
skip_covered = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"